always end shell scripts with 'true'
this ensures that when the script ends with a short-form control flow
statement of the form [expr0] && ... [exprN], and one of [expr0]-[exprN]
evaluated to false, the exit code of the entire shell script _is_ zero,
as was probably intended.

of course, one could add a '|| true' only to these special cases, but
why not just do it everywhere so one doesn't have to remember to do it
then?