redirect git pull stdout to devnull
option -q doesn't seem to get passed down to
git commit. but, git commit reports its summary to
stdout, which is good because we want stderr to
make it through to the terminal if an error does
occur. though, this seems unlikely, largely
because of the sanity checks beforhand, as well as
-X theirs to automatically resolve merge
conflicts.

we still leave on -q because git fetch reports its
progress (which we don't want to show) to stderr
(which we don't want to redirect) if we don't.