remove $lock if it's not a symlink
this fixes a bug that occured where $lock was a
directory.

before the fix, when $lock was a directory, any
symlink attempts with $lock as the link name would
succeed. then, two concurrent instances of update-
dotfiles would each pass the `ln -s $$ $lock`
test, instead of one failing it as intended.