directly rebase onto origin/master
don't try to duplicate the cool "git pull
--rebase" behavior[1] of checking the reflog for
"origin/master" updates and stuff; though
certainly we would like to, and it's probably
possible using git plumbing commands, we can just
promise to not ever force-push to upstream.

it turns out our attempt to duplicate this cool
behavior relies on an assumption which is not
always the case: we assumed HEAD is up-to-date
with $old_om. if users only use this script to
update their dotfiles, that will indeed be the
case, but a user could do a "git fetch"
themselves, updating origin/master, then invoke
this script. then our assumption would not be
correct.

one alternative considered was to use some kind of
"git config dotfiles.currentUpstream" and then
rebase against that and onto origin/master, but
for now, we will decide not to do that. however,
I'm a fickle person so we'll see.

also, keep in mind we don't want to just use "git
pull --rebase," because we want to keep the fetch
and the rebase separate, for two reasons:

- keeping the fetch separate lets us do the nice
  github API thing, to see if upstream has changed
  from what we have, possibly saving the work of
  fetching at all
- keeping the rebase separate lets us check if
  we're already up-to-date before having to make
  sure the working tree is clean, allowing
  no-rebase-needed situations to have dirty
  working trees without erroring

[1]: http://gitolite.com/git-pull--rebase